refactor gc module
