+++ core/common/paddle_tensor_impl.h
@@ -420,4 +420,32 @@ void PaddleTensor<T>::slice(size_t begin_idx, size_t end_idx,
   ret->scaling_factor() = scaling_factor();
 }
 
+template<typename T>
+std::shared_ptr<TensorAdapter<T>> PaddleTensor<T>::operator[](size_t index) {
+    PADDLE_ENFORCE_GT(this->shape().size(), 1,
+                     "lhs's shape must great than 1.");
+    auto slice_shape = this->shape();
+    slice_shape.erase(slice_shape.begin());
+    std::shared_ptr<PaddleTensor<T>> ret = std::make_shared<PaddleTensor<T>>(_device_ctx);
+    ret->reshape(slice_shape);
+
+    this->slice(index, index + 1, ret.get());
+    ret->reshape(slice_shape);
+    return ret;
+}
+
+template<typename T>
+const std::shared_ptr<TensorAdapter<T>> PaddleTensor<T>::operator[](size_t index) const {
+    PADDLE_ENFORCE_GT(this->shape().size(), 1,
+                     "lhs's shape must great than 1.");
+    auto slice_shape = this->shape();
+    slice_shape.erase(slice_shape.begin());
+    std::shared_ptr<PaddleTensor<T>> ret = std::make_shared<PaddleTensor<T>>(_device_ctx);
+    ret->reshape(slice_shape);
+
+    this->slice(index, index + 1, ret.get());
+    ret->reshape(slice_shape);
+    return ret;
+}
+
 } // namespace common
