+++ core/common/tensor_adapter.h
@@ -14,6 +14,7 @@
 
 #pragma once
 
+#include <memory>
 #include <algorithm>
 #include <vector>
 #include <iostream>
@@ -123,6 +124,12 @@ public:
     //      x.slice(1, 2, y)
     //      y.shape = [ 1, 3, 4]
     virtual void slice(size_t begin_idx, size_t end_idx, TensorAdapter* out) const = 0;
+
+    // return point of slice(index, index + 1)
+    virtual std::shared_ptr<TensorAdapter<T>> operator[](size_t index) = 0;
+
+    // return point of slice(index, index + 1)
+    virtual const std::shared_ptr<TensorAdapter<T>> operator[](size_t index) const = 0;
 };
 
 template<typename T>
