+++ core/privc/CMakeLists.txt
@@ -1,6 +1,5 @@
 set(PRIVC_SRCS
     "privc_context.cc"
-    "gc_bit.cc"
     "ot.cc"
 )
 
@@ -14,6 +13,4 @@ target_link_libraries(privc common mpc_tensor)
 cc_test(privc_fixedpoint_tensor_test SRCS fixedpoint_tensor_test.cc DEPS privc)
 cc_test(triplet_generator_test SRCS triplet_generator_test.cc DEPS privc)
 
-cc_test(privc_gc_bit_test SRCS privc_gc_bit_test.cc DEPS privc)
-cc_test(privc_gc_fixedpoint_test SRCS privc_gc_fixedpoint_test.cc DEPS privc)
