+++ core/privc/fixedpoint_tensor_imp.h
@@ -20,7 +20,6 @@
 #include "paddle/fluid/platform/enforce.h"
 #include "../common/prng.h"
 #include "core/common/paddle_tensor.h"
-#include "core/privc/gc_fixedpoint_tensor.h"
 
 namespace privc {
 
@@ -384,161 +383,6 @@ void FixedPointTensor<T, N>::mat_mul_impl(const FixedPointTensor<T, N>* rhs,
                     });
         *(ret_ptr + i) = trunc_sum;
     }
-}
-
-template<typename T, size_t N>
-template<typename T_>
-void FixedPointTensor<T, N>::relu_impl(FixedPointTensor<T, N>* ret,
-                                       const Type2Type<int64_t>) const {
-    PADDLE_ENFORCE_EQ(ret->numel(), numel(), "input numel mot match.");
-    // ac to gc
-    GCFixedPointTensor<N> x(share(), 0);
-    GCFixedPointTensor<N> y(share(), 1);
-    auto gc_shape = get_gc_shape(shape());
-    GCFixedPointTensor<N> gc(gc_shape);
-    x.bitwise_add(&y, &gc);
-    // relu bc
-    auto ret_bc = tensor_factory()->template create<int64_t>(shape());
-    gc.relu_bc(ret_bc.get());
-    // bc to ac
-    to_ac_num(ret_bc.get(), ret->mutable_share());
-
-}
-
-template<typename T, size_t N>
-template<typename T_>
-void FixedPointTensor<T, N>::sigmoid_impl(FixedPointTensor<T, N>* ret,
-                                       const Type2Type<int64_t>) const {
-    PADDLE_ENFORCE_EQ(ret->numel(), numel(), "input numel mot match.");
-    // ac to gc
-    GCFixedPointTensor<N> x(share(), 0);
-    GCFixedPointTensor<N> y(share(), 1);
-    auto gc_shape = get_gc_shape(shape());
-    GCFixedPointTensor<N> gc(gc_shape);
-    x.bitwise_add(&y, &gc);
-    // gc logistic
-    GCFixedPointTensor<N> ret_gc(gc_shape);
-    gc.logistic(&ret_gc);
-    // gc to ac
-    auto bc_shape = gc_shape;
-    bc_shape.erase(bc_shape.begin());
-    bc_shape.erase(bc_shape.begin());
-    auto res_lsb = tensor_factory()->template create<T>(bc_shape);
-    ret_gc.lsb(res_lsb.get());
-    to_ac_num(res_lsb.get(), ret->mutable_share());
-}
-
-template<typename T, size_t N>
-template<typename T_>
-void FixedPointTensor<T, N>::argmax_impl(FixedPointTensor<T, N>* ret,
-                                       const Type2Type<int64_t>) const {
-    PADDLE_ENFORCE_EQ(ret->shape()[1], shape()[1],
-                      "lhs column not match with return column.");
-    PADDLE_ENFORCE_EQ(ret->numel(), numel(),
-                      "input numel mot match with return.");
-    // ac to gc
-    GCFixedPointTensor<N> x(share(), 0);
-    GCFixedPointTensor<N> y(share(), 1);
-    auto gc_shape = get_gc_shape(shape());
-    GCFixedPointTensor<N> gc(gc_shape);
-    x.bitwise_add(&y, &gc);
-    // gc argmax
-    auto ret_gc_shape = get_block_shape(shape());
-    // 1 bit is enough for argmax ret
-    ret_gc_shape.insert(ret_gc_shape.begin(), 1);
-    GCFixedPointTensor<N> ret_gc(ret_gc_shape);
-    GCFixedPointTensor<N>::argmax_one_hot(&gc, &ret_gc);
-    // gc to ac
-    auto ret_ = tensor_factory()->template create<int64_t>(ret->shape());
-    ret_gc.lsb(ret_.get());
-    to_ac_num(ret_.get(), ret_.get());
-    // to fixedpoint number
-    std::transform(ret_->data(), ret_->data() + ret_->numel(),
-                   ret->mutable_share()->data(),
-                   [] (T a) {
-                       // int to fixedpoint
-                       return (a << N); 
-                    });
-}
-
-template<typename T, size_t N>
-template<typename T_>
-void FixedPointTensor<T, N>::long_div_impl(const FixedPointTensor<T, N>* rhs,
-                                       FixedPointTensor<T, N>* ret,
-                                       const Type2Type<int64_t>) const {
-    PADDLE_ENFORCE_EQ(ret->numel(), numel(),
-            "input of lhs's numel no match with return.");
-    PADDLE_ENFORCE_EQ(ret->numel(), rhs->numel(),
-            "input of rhs's numel no match with return.");
-    // ac to gc
-    GCFixedPointTensor<N> l_x(share(), 0);
-    GCFixedPointTensor<N> l_y(share(), 1);
-
-    GCFixedPointTensor<N> r_x(rhs->share(), 0);
-    GCFixedPointTensor<N> r_y(rhs->share(), 1);
-
-    auto gc_shape = get_gc_shape(shape());
-    GCFixedPointTensor<N> l_gc(gc_shape);
-    l_x.bitwise_add(&l_y, &l_gc);
-
-    GCFixedPointTensor<N> r_gc(gc_shape);
-    r_x.bitwise_add(&r_y, &r_gc);
-
-    // gc logistic
-    GCFixedPointTensor<N> ret_gc(gc_shape);
-    l_gc.bitwise_div(&r_gc, &ret_gc);
-    // gc to ac
-    auto bc_shape = gc_shape;
-    bc_shape.erase(bc_shape.begin());
-    bc_shape.erase(bc_shape.begin());
-    auto res_lsb = tensor_factory()->template create<T>(bc_shape);
-    ret_gc.lsb(res_lsb.get());
-    to_ac_num(res_lsb.get(), ret->mutable_share());
-}
-
-// reduce last dim
-template <typename T, size_t N>
-void FixedPointTensor<T, N>::reduce(FixedPointTensor<T, N>* ret) const {
-    std::vector<size_t> shape_reduce = ret->shape();
-    auto shape = this->shape();
-    size_t numel = ret->numel();
-    for (int i = 0; i < this->numel(); ++i) {
-        T* ret_ptr = ret->mutable_share()->data() + i;
-        *ret_ptr = 0;
-        std::for_each(share()->data() + i * shape[shape.size() - 1],
-                      share()->data() + (i + 1) * shape[shape.size() - 1],
-                      [ret_ptr](T a) { *ret_ptr += a; });
-    }
-}
-
-template<typename T, size_t N>
-template<typename T_>
-void FixedPointTensor<T, N>::softmax_impl(FixedPointTensor<T, N>* ret,
-                                     bool use_relu,
-                                     const Type2Type<int64_t>) const {
-    auto tmp = tensor_factory()->template create<int64_t>(shape());
-    FixedPointTensor<T, N> x(tmp.get());
-    if (use_relu) {
-        this->relu(&x);
-    } else {
-        this->exp(&x);
-    }
-
-    auto tmp1 = tensor_factory()->template create<int64_t>({shape()[1]});
-    FixedPointTensor<T, N> sum(tmp1.get());
-    x.reduce(&sum);
-
-    auto sum_ext = tensor_factory()->template create<int64_t>(shape());
-    const T* sum_ptr = sum.share()->data();
-    for (int i = 0; i < shape()[0]; ++i) {
-        T* col_ptr = sum_ext->data() + i * shape()[1];
-        //std::transform(sum_ptr, sum_ptr + shape()[0], col_ptr,
-        //               [](T a) { return a; });
-        std::for_each(col_ptr, col_ptr + shape()[1],
-                       [sum_ptr, i]( T& a) { a = *(sum_ptr + i); });
-    }
-    FixedPointTensor<T, N> sum_ext_fixed(sum_ext.get());
-    x.long_div(&sum_ext_fixed, ret);
 }
 
 } // namespace privc
